added queries + fixed test coverage issues + lots of optimizations
