--- contracts/liquidity_book/lb_pair/src/unittest/query_test.rs
@@ -1,1 +1,0 @@
-
