+++ contracts/liquidity_book/staking_contract/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod contract;
+pub mod handles;
 pub mod helper;
 pub mod msg;
 pub mod query;
