+++ contracts/liquidity_book/tests/Cargo.toml
@@ -38,4 +38,5 @@ shade-multi-test = { path = "../../../packages/multi_test", features = [
   "router",
   "lb_factory"
 ] }
+lb_pair ={ path = "../lb_pair"}
 serial_test = "2.0.0"
