+++ contracts/liquidity_book/tests/src/multitests/lb_pair_liquidity.rs
@@ -13,7 +13,10 @@ use shade_protocol::{
     liquidity_book::lb_pair::RemoveLiquidity,
     multi_test::App,
 };
-use std::{cmp::Ordering, ops::Add};
+use std::{
+    cmp::Ordering,
+    ops::{Add, Sub},
+};
 
 use crate::multitests::test_helper::*;
 
@@ -1280,6 +1283,744 @@ pub fn test_revert_burn_zero() -> Result<(), anyhow::Error> {
     assert_eq!(
         res.unwrap_err(),
         StdError::generic_err(format!("Zero Shares for bin id: {:?}", ACTIVE_ID))
+    );
+
+    Ok(())
+}
+
+#[test]
+#[serial]
+pub fn test_revert_on_deadline() -> Result<(), anyhow::Error> {
+    let addrs = init_addrs();
+    let (mut app, _lb_factory, deployed_contracts, lb_pair, _lb_token) = lb_pair_setup()?;
+
+    let amount_x = Uint128::from(600 * 100_000_000_u128); //10^8
+    let amount_y = Uint128::from(100 * 100_000_000_u128);
+    let nb_bins_x = 6u32;
+    let nb_bins_y = 6u32;
+
+    let token_x = extract_contract_info(&deployed_contracts, SILK)?;
+    let token_y = extract_contract_info(&deployed_contracts, SHADE)?;
+
+    let tokens_to_mint = vec![(SILK, amount_x), (SHADE, amount_y)];
+
+    mint_token_helper(
+        &mut app,
+        &deployed_contracts,
+        &addrs,
+        addrs.batman().into_string(),
+        tokens_to_mint.clone(),
+    )?;
+
+    increase_allowance_helper(
+        &mut app,
+        &deployed_contracts,
+        addrs.batman().into_string(),
+        lb_pair.lb_pair.contract.address.to_string(),
+        tokens_to_mint,
+    )?;
+
+    //Adding liquidity
+    let mut liquidity_parameters = liquidity_parameters_generator(
+        &deployed_contracts,
+        ACTIVE_ID,
+        token_x,
+        token_y,
+        amount_x,
+        amount_y,
+        nb_bins_x,
+        nb_bins_y,
+    )?;
+
+    let current_time = app.block_info().time.seconds();
+    let deadline = current_time.sub(1);
+
+    liquidity_parameters.deadline = deadline;
+
+    let res = lb_pair::add_liquidity(
+        &mut app,
+        addrs.batman().as_str(),
+        &lb_pair.lb_pair.contract,
+        liquidity_parameters,
+    );
+
+    assert_eq!(
+        res.unwrap_err(),
+        StdError::generic_err(format!(
+            "Deadline exceeded. Deadline: {deadline}, Current timestamp: {current_time}"
+        ))
+    );
+
+    Ok(())
+}
+
+#[test]
+#[serial]
+pub fn test_revert_on_wrong_pair() -> Result<(), anyhow::Error> {
+    let addrs = init_addrs();
+    let (mut app, _lb_factory, deployed_contracts, lb_pair, _lb_token) = lb_pair_setup()?;
+
+    let amount_x = Uint128::from(600 * 100_000_000_u128); //10^8
+    let amount_y = Uint128::from(100 * 100_000_000_u128);
+    let nb_bins_x = 6u32;
+    let nb_bins_y = 6u32;
+
+    let token_x = extract_contract_info(&deployed_contracts, SSCRT)?;
+    let token_y = extract_contract_info(&deployed_contracts, SHADE)?;
+
+    let tokens_to_mint = vec![(SSCRT, amount_x), (SHADE, amount_y)];
+
+    mint_token_helper(
+        &mut app,
+        &deployed_contracts,
+        &addrs,
+        addrs.batman().into_string(),
+        tokens_to_mint.clone(),
+    )?;
+
+    increase_allowance_helper(
+        &mut app,
+        &deployed_contracts,
+        addrs.batman().into_string(),
+        lb_pair.lb_pair.contract.address.to_string(),
+        tokens_to_mint,
+    )?;
+
+    //Adding liquidity
+    let liquidity_parameters = liquidity_parameters_generator(
+        &deployed_contracts,
+        ACTIVE_ID,
+        token_x,
+        token_y,
+        amount_x,
+        amount_y,
+        nb_bins_x,
+        nb_bins_y,
+    )?;
+
+    let res = lb_pair::add_liquidity(
+        &mut app,
+        addrs.batman().as_str(),
+        &lb_pair.lb_pair.contract,
+        liquidity_parameters,
+    );
+
+    assert_eq!(
+        res.unwrap_err(),
+        StdError::generic_err(format!("Wrong Pair"))
+    );
+
+    Ok(())
+}
+
+#[test]
+#[serial]
+pub fn test_revert_on_amount_slippage() -> Result<(), anyhow::Error> {
+    let addrs = init_addrs();
+    let (mut app, _lb_factory, deployed_contracts, lb_pair, _lb_token) = lb_pair_setup()?;
+
+    let amount_x = Uint128::from(600 * 100_000_000_u128); //10^8
+    let amount_y = Uint128::from(100 * 100_000_000_u128);
+    let nb_bins_x = 6u32;
+    let nb_bins_y = 6u32;
+
+    let token_x = extract_contract_info(&deployed_contracts, SILK)?;
+    let token_y = extract_contract_info(&deployed_contracts, SHADE)?;
+
+    let tokens_to_mint = vec![(SILK, amount_x), (SHADE, amount_y)];
+
+    mint_token_helper(
+        &mut app,
+        &deployed_contracts,
+        &addrs,
+        addrs.batman().into_string(),
+        tokens_to_mint.clone(),
+    )?;
+
+    increase_allowance_helper(
+        &mut app,
+        &deployed_contracts,
+        addrs.batman().into_string(),
+        lb_pair.lb_pair.contract.address.to_string(),
+        tokens_to_mint,
+    )?;
+
+    //Adding liquidity
+    let mut liquidity_parameters = liquidity_parameters_generator(
+        &deployed_contracts,
+        ACTIVE_ID,
+        token_x,
+        token_y,
+        amount_x,
+        amount_y,
+        nb_bins_x,
+        nb_bins_y,
+    )?;
+
+    liquidity_parameters.amount_x_min = Uint128::MAX;
+    liquidity_parameters.amount_y_min = Uint128::MAX;
+
+    let res = lb_pair::add_liquidity(
+        &mut app,
+        addrs.batman().as_str(),
+        &lb_pair.lb_pair.contract,
+        liquidity_parameters.clone(),
+    );
+
+    assert_eq!(
+        res.unwrap_err(),
+        StdError::generic_err(format!(
+            "Amount slippage caught. AmountXMin: {0}, AmountX: {1}, AmountYMin: {2}, AmountY: {3}",
+            liquidity_parameters.amount_x_min,
+            amount_x,
+            liquidity_parameters.amount_y_min,
+            amount_y
+        ))
+    );
+    Ok(())
+}
+
+#[test]
+#[serial]
+pub fn test_revert_on_length_mismatch() -> Result<(), anyhow::Error> {
+    let addrs = init_addrs();
+    let (mut app, _lb_factory, deployed_contracts, lb_pair, _lb_token) = lb_pair_setup()?;
+
+    let amount_x = Uint128::from(600 * 100_000_000_u128); //10^8
+    let amount_y = Uint128::from(100 * 100_000_000_u128);
+    let nb_bins_x = 6u32;
+    let nb_bins_y = 6u32;
+
+    let token_x = extract_contract_info(&deployed_contracts, SILK)?;
+    let token_y = extract_contract_info(&deployed_contracts, SHADE)?;
+
+    let tokens_to_mint = vec![(SILK, amount_x), (SHADE, amount_y)];
+
+    mint_token_helper(
+        &mut app,
+        &deployed_contracts,
+        &addrs,
+        addrs.batman().into_string(),
+        tokens_to_mint.clone(),
+    )?;
+
+    increase_allowance_helper(
+        &mut app,
+        &deployed_contracts,
+        addrs.batman().into_string(),
+        lb_pair.lb_pair.contract.address.to_string(),
+        tokens_to_mint,
+    )?;
+
+    //Adding liquidity
+    let mut liquidity_parameters = liquidity_parameters_generator(
+        &deployed_contracts,
+        ACTIVE_ID,
+        token_x,
+        token_y,
+        amount_x,
+        amount_y,
+        nb_bins_x,
+        nb_bins_y,
+    )?;
+
+    liquidity_parameters.delta_ids.pop();
+
+    let res = lb_pair::add_liquidity(
+        &mut app,
+        addrs.batman().as_str(),
+        &lb_pair.lb_pair.contract,
+        liquidity_parameters.clone(),
+    );
+
+    assert_eq!(
+        res.unwrap_err(),
+        StdError::generic_err(format!("Lengths mismatch"))
+    );
+    Ok(())
+}
+
+#[test]
+#[serial]
+pub fn test_revert_on_id_desired_overflow() -> Result<(), anyhow::Error> {
+    let addrs = init_addrs();
+    let (mut app, _lb_factory, deployed_contracts, lb_pair, _lb_token) = lb_pair_setup()?;
+
+    let amount_x = Uint128::from(600 * 100_000_000_u128); //10^8
+    let amount_y = Uint128::from(100 * 100_000_000_u128);
+    let nb_bins_x = 6u32;
+    let nb_bins_y = 6u32;
+
+    let token_x = extract_contract_info(&deployed_contracts, SILK)?;
+    let token_y = extract_contract_info(&deployed_contracts, SHADE)?;
+
+    let tokens_to_mint = vec![(SILK, amount_x), (SHADE, amount_y)];
+
+    mint_token_helper(
+        &mut app,
+        &deployed_contracts,
+        &addrs,
+        addrs.batman().into_string(),
+        tokens_to_mint.clone(),
+    )?;
+
+    increase_allowance_helper(
+        &mut app,
+        &deployed_contracts,
+        addrs.batman().into_string(),
+        lb_pair.lb_pair.contract.address.to_string(),
+        tokens_to_mint,
+    )?;
+
+    //Adding liquidity
+    let mut liquidity_parameters = liquidity_parameters_generator(
+        &deployed_contracts,
+        ACTIVE_ID,
+        token_x,
+        token_y,
+        amount_x,
+        amount_y,
+        nb_bins_x,
+        nb_bins_y,
+    )?;
+
+    liquidity_parameters.active_id_desired = U24::MAX + 1;
+
+    let res = lb_pair::add_liquidity(
+        &mut app,
+        addrs.batman().as_str(),
+        &lb_pair.lb_pair.contract,
+        liquidity_parameters.clone(),
+    );
+
+    assert_eq!(
+        res.unwrap_err(),
+        StdError::generic_err(format!(
+            "Id desired overflows. Id desired: {0}, Id slippage: {1}",
+            liquidity_parameters.active_id_desired, 15
+        ))
+    );
+    Ok(())
+}
+
+#[test]
+#[serial]
+pub fn test_revert_on_id_slippage_caught() -> Result<(), anyhow::Error> {
+    let addrs = init_addrs();
+    let (mut app, _lb_factory, deployed_contracts, lb_pair, _lb_token) = lb_pair_setup()?;
+
+    let amount_x = Uint128::from(600 * 100_000_000_u128); //10^8
+    let amount_y = Uint128::from(100 * 100_000_000_u128);
+    let nb_bins_x = 6u32;
+    let nb_bins_y = 6u32;
+
+    let token_x = extract_contract_info(&deployed_contracts, SILK)?;
+    let token_y = extract_contract_info(&deployed_contracts, SHADE)?;
+
+    let tokens_to_mint = vec![(SILK, amount_x), (SHADE, amount_y)];
+
+    mint_token_helper(
+        &mut app,
+        &deployed_contracts,
+        &addrs,
+        addrs.batman().into_string(),
+        tokens_to_mint.clone(),
+    )?;
+
+    increase_allowance_helper(
+        &mut app,
+        &deployed_contracts,
+        addrs.batman().into_string(),
+        lb_pair.lb_pair.contract.address.to_string(),
+        tokens_to_mint,
+    )?;
+
+    //Adding liquidity
+    let mut liquidity_parameters = liquidity_parameters_generator(
+        &deployed_contracts,
+        ACTIVE_ID,
+        token_x,
+        token_y,
+        amount_x,
+        amount_y,
+        nb_bins_x,
+        nb_bins_y,
+    )?;
+
+    liquidity_parameters.active_id_desired = ACTIVE_ID + 100;
+    liquidity_parameters.id_slippage = 10;
+
+    let res = lb_pair::add_liquidity(
+        &mut app,
+        addrs.batman().as_str(),
+        &lb_pair.lb_pair.contract,
+        liquidity_parameters.clone(),
+    );
+
+    assert_eq!(
+        res.unwrap_err(),
+        StdError::generic_err(format!(
+            "Id slippage caught. Active id desired: {0}, Id slippage: {1}, Active id: {2}",
+            liquidity_parameters.active_id_desired, 10, ACTIVE_ID
+        ))
+    );
+    Ok(())
+}
+
+#[test]
+#[serial]
+pub fn test_revert_on_delta_ids_overflow() -> Result<(), anyhow::Error> {
+    let addrs = init_addrs();
+    let (mut app, _lb_factory, deployed_contracts, lb_pair, _lb_token) = lb_pair_setup()?;
+
+    let amount_x = Uint128::from(600 * 100_000_000_u128); //10^8
+    let amount_y = Uint128::from(100 * 100_000_000_u128);
+    let nb_bins_x = 6u32;
+    let nb_bins_y = 6u32;
+
+    let token_x = extract_contract_info(&deployed_contracts, SILK)?;
+    let token_y = extract_contract_info(&deployed_contracts, SHADE)?;
+
+    let tokens_to_mint = vec![(SILK, amount_x), (SHADE, amount_y)];
+
+    mint_token_helper(
+        &mut app,
+        &deployed_contracts,
+        &addrs,
+        addrs.batman().into_string(),
+        tokens_to_mint.clone(),
+    )?;
+
+    increase_allowance_helper(
+        &mut app,
+        &deployed_contracts,
+        addrs.batman().into_string(),
+        lb_pair.lb_pair.contract.address.to_string(),
+        tokens_to_mint,
+    )?;
+
+    //Adding liquidity
+    let mut liquidity_parameters = liquidity_parameters_generator(
+        &deployed_contracts,
+        ACTIVE_ID,
+        token_x,
+        token_y,
+        amount_x,
+        amount_y,
+        nb_bins_x,
+        nb_bins_y,
+    )?;
+
+    liquidity_parameters.delta_ids[0] = liquidity_parameters.delta_ids[0] - ACTIVE_ID as i64;
+
+    let res = lb_pair::add_liquidity(
+        &mut app,
+        addrs.batman().as_str(),
+        &lb_pair.lb_pair.contract,
+        liquidity_parameters.clone(),
+    );
+
+    assert_eq!(
+        res.unwrap_err(),
+        StdError::generic_err(format!(
+            "Delta id overflows. Delta Id: {0}",
+            liquidity_parameters.delta_ids[0]
+        ))
+    );
+    Ok(())
+}
+
+#[test]
+#[serial]
+pub fn test_revert_on_empty_liquidity_config() -> Result<(), anyhow::Error> {
+    let addrs = init_addrs();
+    let (mut app, _lb_factory, deployed_contracts, lb_pair, _lb_token) = lb_pair_setup()?;
+
+    let amount_x = Uint128::from(600 * 100_000_000_u128); //10^8
+    let amount_y = Uint128::from(100 * 100_000_000_u128);
+    let nb_bins_x = 6u32;
+    let nb_bins_y = 6u32;
+
+    let token_x = extract_contract_info(&deployed_contracts, SILK)?;
+    let token_y = extract_contract_info(&deployed_contracts, SHADE)?;
+
+    let tokens_to_mint = vec![(SILK, amount_x), (SHADE, amount_y)];
+
+    mint_token_helper(
+        &mut app,
+        &deployed_contracts,
+        &addrs,
+        addrs.batman().into_string(),
+        tokens_to_mint.clone(),
+    )?;
+
+    increase_allowance_helper(
+        &mut app,
+        &deployed_contracts,
+        addrs.batman().into_string(),
+        lb_pair.lb_pair.contract.address.to_string(),
+        tokens_to_mint,
+    )?;
+
+    //Adding liquidity
+    let mut liquidity_parameters = liquidity_parameters_generator(
+        &deployed_contracts,
+        ACTIVE_ID,
+        token_x,
+        token_y,
+        amount_x,
+        amount_y,
+        nb_bins_x,
+        nb_bins_y,
+    )?;
+
+    liquidity_parameters.delta_ids = vec![];
+    liquidity_parameters.distribution_x = vec![];
+    liquidity_parameters.distribution_y = vec![];
+
+    let res = lb_pair::add_liquidity(
+        &mut app,
+        addrs.batman().as_str(),
+        &lb_pair.lb_pair.contract,
+        liquidity_parameters.clone(),
+    );
+
+    assert_eq!(
+        res.unwrap_err(),
+        StdError::generic_err(format!("Empty Market Configuration"))
+    );
+    Ok(())
+}
+
+#[test]
+#[serial]
+pub fn testing_implicit_swap() -> Result<(), anyhow::Error> {
+    let addrs = init_addrs();
+    let (mut app, _lb_factory, deployed_contracts, lb_pair, _lb_token) = lb_pair_setup()?;
+
+    let amount_x = Uint128::from(5000000000000000000000u128); //10^8
+    let amount_y = Uint128::from(1000000000000000000000u128);
+    let nb_bins_x = 6u32;
+    let nb_bins_y = 6u32;
+
+    let token_x = extract_contract_info(&deployed_contracts, SILK)?;
+    let token_y = extract_contract_info(&deployed_contracts, SHADE)?;
+
+    let tokens_to_mint = vec![(SILK, amount_x), (SHADE, amount_y)];
+
+    mint_token_helper(
+        &mut app,
+        &deployed_contracts,
+        &addrs,
+        addrs.batman().into_string(),
+        tokens_to_mint.clone(),
+    )?;
+
+    increase_allowance_helper(
+        &mut app,
+        &deployed_contracts,
+        addrs.batman().into_string(),
+        lb_pair.lb_pair.contract.address.to_string(),
+        tokens_to_mint,
+    )?;
+
+    //Adding liquidity
+    let liquidity_parameters = liquidity_parameters_generator(
+        &deployed_contracts,
+        ACTIVE_ID,
+        token_x,
+        token_y,
+        amount_x,
+        amount_y,
+        nb_bins_x,
+        nb_bins_y,
+    )?;
+
+    lb_pair::add_liquidity(
+        &mut app,
+        addrs.batman().as_str(),
+        &lb_pair.lb_pair.contract,
+        liquidity_parameters.clone(),
+    )?;
+
+    //REPEAT WITH EXTREME difference in ratio
+    let amount_x = Uint128::from(500000000000000000000u128); //10^8
+    let amount_y = Uint128::from(500000000000000000000u128);
+    let nb_bins_x = 6u32;
+    let nb_bins_y = 6u32;
+
+    let token_x = extract_contract_info(&deployed_contracts, SILK)?;
+    let token_y = extract_contract_info(&deployed_contracts, SHADE)?;
+
+    let tokens_to_mint = vec![(SILK, amount_x), (SHADE, amount_y)];
+
+    mint_token_helper(
+        &mut app,
+        &deployed_contracts,
+        &addrs,
+        addrs.batman().into_string(),
+        tokens_to_mint.clone(),
+    )?;
+
+    increase_allowance_helper(
+        &mut app,
+        &deployed_contracts,
+        addrs.batman().into_string(),
+        lb_pair.lb_pair.contract.address.to_string(),
+        tokens_to_mint,
+    )?;
+
+    //Adding liquidity
+    let liquidity_parameters = liquidity_parameters_generator(
+        &deployed_contracts,
+        ACTIVE_ID,
+        token_x,
+        token_y,
+        amount_x,
+        amount_y,
+        nb_bins_x,
+        nb_bins_y,
+    )?;
+
+    let (protocol_fee_x, protocol_fee_y) =
+        lb_pair::query_protocol_fees(&app, &lb_pair.lb_pair.contract)?;
+
+    assert_eq!(0, protocol_fee_x);
+    assert_eq!(0, protocol_fee_y);
+
+    lb_pair::add_liquidity(
+        &mut app,
+        addrs.batman().as_str(),
+        &lb_pair.lb_pair.contract,
+        liquidity_parameters.clone(),
+    )?;
+
+    let (protocol_fee_x, protocol_fee_y) =
+        lb_pair::query_protocol_fees(&app, &lb_pair.lb_pair.contract)?;
+
+    assert_eq!(0, protocol_fee_x);
+    assert!(protocol_fee_y > 0);
+
+    Ok(())
+}
+
+#[test]
+#[serial]
+pub fn test_revert_burn_on_wrong_pair() -> Result<(), anyhow::Error> {
+    let addrs = init_addrs();
+    let (mut app, _lb_factory, deployed_contracts, lb_pair, lb_token) = lb_pair_setup()?;
+    let amount_x = Uint128::from(600 * 100_000_000_u128); //10^8
+    let amount_y = Uint128::from(100 * 100_000_000_u128);
+    let nb_bins_x = 6;
+    let nb_bins_y = 6;
+
+    let token_x = extract_contract_info(&deployed_contracts, SILK)?;
+    let token_y = extract_contract_info(&deployed_contracts, SHADE)?;
+
+    let tokens_to_mint = vec![(SILK, amount_x), (SHADE, amount_y)];
+
+    mint_token_helper(
+        &mut app,
+        &deployed_contracts,
+        &addrs,
+        addrs.batman().into_string(),
+        tokens_to_mint.clone(),
+    )?;
+
+    increase_allowance_helper(
+        &mut app,
+        &deployed_contracts,
+        addrs.batman().into_string(),
+        lb_pair.lb_pair.contract.address.to_string(),
+        tokens_to_mint,
+    )?;
+
+    // Adding liquidity with nb_bins_x and nb_bins_y
+    let liquidity_parameters = liquidity_parameters_generator(
+        &deployed_contracts,
+        ACTIVE_ID,
+        token_x.clone(),
+        token_y.clone(),
+        amount_x,
+        amount_y,
+        nb_bins_x,
+        nb_bins_y,
+    )?;
+
+    lb_pair::add_liquidity(
+        &mut app,
+        addrs.batman().as_str(),
+        &lb_pair.lb_pair.contract,
+        liquidity_parameters,
+    )?;
+
+    let total_bins = get_total_bins(nb_bins_x as u32, nb_bins_y as u32) as u32;
+    let mut balances = vec![Uint256::zero(); total_bins as usize];
+    let mut ids = vec![0u32; total_bins as usize];
+
+    for i in 0..total_bins {
+        let id = get_id(ACTIVE_ID, i, nb_bins_y);
+        ids[i as usize] = id;
+        balances[i as usize] = lb_token::query_balance(
+            &app,
+            &lb_token,
+            addrs.batman(),
+            addrs.batman(),
+            String::from("viewing_key"),
+            id.to_string(),
+        )?;
+    }
+
+    let (reserves_x, reserves_y) = lb_pair::query_reserves(&app, &lb_pair.lb_pair.contract)?;
+
+    //SWAPPED tokens order
+    let token_x = extract_contract_info(&deployed_contracts, SSCRT)?;
+    let token_y = extract_contract_info(&deployed_contracts, SHADE)?;
+
+    let res = lb_pair::remove_liquidity(
+        &mut app,
+        addrs.batman().as_str(),
+        &lb_pair.lb_pair.contract,
+        RemoveLiquidity {
+            token_x: token_type_snip20_generator(&token_x)?,
+            token_y: token_type_snip20_generator(&token_y)?,
+            bin_step: lb_pair.bin_step,
+            amount_x_min: Uint128::from(reserves_x),
+            amount_y_min: Uint128::from(reserves_y),
+            ids: ids.clone(),
+            amounts: balances.clone(),
+            deadline: 99999999999,
+        },
+    );
+    assert_eq!(
+        res.unwrap_err(),
+        StdError::generic_err(format!("Wrong Pair"))
+    );
+
+    //WRONG order
+
+    //SWAPPED tokens order
+    let token_x = extract_contract_info(&deployed_contracts, SILK)?;
+    let token_y = extract_contract_info(&deployed_contracts, SSCRT)?;
+
+    let res = lb_pair::remove_liquidity(
+        &mut app,
+        addrs.batman().as_str(),
+        &lb_pair.lb_pair.contract,
+        RemoveLiquidity {
+            token_x: token_type_snip20_generator(&token_x)?,
+            token_y: token_type_snip20_generator(&token_y)?,
+            bin_step: lb_pair.bin_step,
+            amount_x_min: Uint128::from(reserves_x),
+            amount_y_min: Uint128::from(reserves_y),
+            ids,
+            amounts: balances,
+            deadline: 99999999999,
+        },
+    );
+    assert_eq!(
+        res.unwrap_err(),
+        StdError::generic_err(format!("Wrong Pair"))
     );
 
     Ok(())
