+++ packages/shade_protocol/src/contract_interfaces/liquidity_book/lb_libraries/bin_helper.rs
@@ -244,18 +244,11 @@ impl BinHelper {
 
         let (amount_x, amount_y) = amounts_in.decode();
 
-        // let (bin_reserves_x, bin_reserves_y) = bin_reserves.decode();
-
         let (received_amount_x, received_amount_y) = Self::get_amount_out_of_bin(
             bin_reserves.add(amounts_in),
             shares,
             total_supply + shares,
         )?;
-
-        // println!(
-        //     "received_amount_x {:?}\nreceived_amount_y:{:?}\namount_x {:?}\namount_y {:?}\nshare {:?}\ntotal_supply {:?}\nbin_reserves_x {:?}\nbin_reserves_y {:?}",
-        //     received_amount_x, received_amount_y,amount_x,amount_y,shares,total_supply,bin_reserves_x,bin_reserves_y
-        // );
 
         let mut fees = Bytes32::default();
