+++ packages/shade_protocol/src/contract_interfaces/liquidity_book/test_coverage.png
